Make `Dates` a weakdep
This exploits that `Compat` only really needs `Dates` (to extend it) on
Julia versions without weakdep support (i.e. before 1.9). So on newer
Julia versions, no (unnecessary) dependence on `Dates` is introduced
without loosing anything on older Julia versions.

Should a post-1.9 feature of `Dates` (or any other stdlib, actually) be
added to `Compat` in the future, the way to do so is probably via a
package extension.